[+] 增加游客解析模式 [+] 增加游客是否使用下载代理服务器
