+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 2025
 
+### 04-27 v2.2.2
+
+- [+] 增加游客解析模式
+- [+] 增加游客是否使用下载代理服务器
+
 ### 04-15 v2.2.1
 
 - [*] 修复 Aira2 推送路径错误
