+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hklist-frontend",
-  "version": "2.2.1",
+  "version": "2.2.2",
   "type": "module",
   "scripts": {
     "dev": "vite",
